+++ src/platform.test.ts
@@ -219,7 +219,7 @@ describe('TestPlatform', () => {
 
     await dynamicPlatform.onStart('Test reason');
     expect(mockLog.info).toHaveBeenCalledWith('onStart called with reason:', 'Test reason');
-    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(47);
+    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(56);
   }, 60000);
 
   it('should start the server', async () => {